requirements.txt
firebase_admin==6.6.0
python-dotenv==1.0.1
pytz==2023.3.post1
redis==4.6.0
redis_om==0.3.3
